+++ NEWS.md
@@ -1,6 +1,14 @@
-# MsBackendMetabolomicsWorkbench 0.1.0
+# MsBackendMetabolomicsWorkbench 0.1
+
+## Changes in 0.1.1
+
+- Possibility to download the zip file of the experiment from the FTP server.
+- Update documentation.
+- Update unit tests.
+
+## Changes in 0.1.0
 
 - Initial version, add the basic Metabolomics Workbench functions.
-- Add functionality to download and cache MS data files from Metabolomics 
+- Add functionality to download and cache MS data files from Metabolomics
   Workbench.
 - Add unit tests.